Refactor: split out UI and strategy
We want to have multiple UIs and multiple strategies!
