+++ gollm/gemini.go
@@ -70,7 +70,7 @@ func (c *GeminiClient) ListModels(ctx context.Context) (modelNames []string, err
 		m, err := models.Next()
 		if err != nil {
 			if err == iterator.Done {
-				return modelNames, nil
+				break
 			}
 			return nil, err
 		}
@@ -103,6 +103,10 @@ func (c *GeminiClient) GenerateCompletion(ctx context.Context, request *Completi
 	geminiResponse, err := model.GenerateContent(ctx, geminiParts...)
 	if err != nil {
 		return nil, err
+	}
+
+	if len(geminiResponse.Candidates) == 0 {
+		return nil, fmt.Errorf("got no responses from gemini")
 	}
 
 	if len(geminiResponse.Candidates) > 1 {
