+++ pkg/llmstrategy/react/strategy.go
@@ -1,0 +1,260 @@
+package react
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"os"
+	"strings"
+	"time"
+
+	"github.com/GoogleCloudPlatform/kubectl-ai/gollm"
+	"github.com/GoogleCloudPlatform/kubectl-ai/pkg/journal"
+	"github.com/GoogleCloudPlatform/kubectl-ai/pkg/ui"
+	"k8s.io/klog/v2"
+)
+
+type Strategy struct {
+	// TemplateFile allows specifying a custom template file
+	TemplateFile string
+
+	// Recorder captures events for diagnostics
+	Recorder journal.Recorder
+
+	RemoveWorkDir bool
+
+	ContentGenerator gollm.Client
+
+	Messages         []Message
+	MaxIterations    int
+	CurrentIteration int
+
+	Query       string
+	PastQueries string
+
+	Kubeconfig string
+
+	Tools map[string]func(input string, kubeconfig string, workDir string) (string, error)
+}
+
+type Message struct {
+	Role    string `json:"role"`
+	Content string `json:"content"`
+}
+
+type ReActResponse struct {
+	Thought string  `json:"thought"`
+	Answer  string  `json:"answer,omitempty"`
+	Action  *Action `json:"action,omitempty"`
+}
+
+type Action struct {
+	Name   string `json:"name"`
+	Reason string `json:"reason"`
+	Input  string `json:"input"`
+}
+
+// Data represents the structure of the data to be filled into the template.
+type Data struct {
+	Query       string
+	PastQueries string
+	History     string
+	Tools       string
+}
+
+// AskLLM asks the LLM for the next action, sending a prompt including the .History
+func (a *Strategy) AskLLM(ctx context.Context) (*ReActResponse, error) {
+	log := klog.FromContext(ctx)
+	log.Info("Asking LLM...")
+
+	data := Data{
+		Query:       a.Query,
+		PastQueries: a.PastQueries,
+		History:     a.History(),
+		Tools:       "kubectl, gcrane, bash",
+	}
+
+	prompt, err := a.generateFromTemplate(data)
+	if err != nil {
+		log.Error(err, "generating from template")
+		return nil, err
+	}
+
+	log.Info("Thinking...", "prompt", prompt)
+
+	response, err := a.ContentGenerator.GenerateCompletion(ctx, &gollm.CompletionRequest{
+		Prompt: prompt,
+	})
+	if err != nil {
+		return nil, fmt.Errorf("generating LLM completion: %w", err)
+	}
+
+	a.Recorder.Write(ctx, &journal.Event{
+		Timestamp: time.Now(),
+		Action:    "llm-response",
+		Payload:   response,
+	})
+
+	reActResp, err := parseReActResponse(response.Response())
+	if err != nil {
+		return nil, fmt.Errorf("parsing ReAct response: %w", err)
+	}
+	return reActResp, nil
+}
+
+func (a *Strategy) History() string {
+	var history strings.Builder
+	for _, msg := range a.Messages {
+		history.WriteString(fmt.Sprintf("%s: %s\n", msg.Role, msg.Content))
+	}
+	return history.String()
+}
+
+func (a *Strategy) RunOnce(ctx context.Context, u ui.UI) error {
+	log := klog.FromContext(ctx)
+	log.Info("Executing query:", "query", a.Query)
+
+	// Create a temporary working directory
+	workDir, err := os.MkdirTemp("", "agent-workdir-*")
+	if err != nil {
+		log.Error(err, "Failed to create temporary working directory")
+		return err
+	}
+	if a.RemoveWorkDir {
+		defer os.RemoveAll(workDir)
+	}
+	log.Info("Created temporary working directory", "workDir", workDir)
+
+	// Main execution loop
+	for a.CurrentIteration < a.MaxIterations {
+		log.Info("Starting iteration", "iteration", a.CurrentIteration)
+
+		// Get next action from LLM
+		reActResp, err := a.AskLLM(ctx)
+		if err != nil {
+			log.Error(err, "Error asking LLM")
+			u.RenderOutput(ctx, fmt.Sprintf("\nSorry, Couldn't complete the task. LLM error %v\n", err), ui.Foreground(ui.ColorRed))
+			return err
+		}
+
+		// Log the thought process
+		log.Info("Thinking...", "thought", reActResp.Thought)
+		a.addMessage(ctx, "assistant", fmt.Sprintf("Thought: %s", reActResp.Thought))
+
+		// Handle final answer
+		if reActResp.Answer != "" {
+			log.Info("Final answer received", "answer", reActResp.Answer)
+			a.addMessage(ctx, "assistant", fmt.Sprintf("Final Answer: %s", reActResp.Answer))
+			u.RenderOutput(ctx, reActResp.Answer, ui.RenderMarkdown())
+			return nil
+		}
+
+		// Handle action
+		if reActResp.Action != nil {
+			log.Info("Executing action",
+				"name", reActResp.Action.Name,
+				"reason", reActResp.Action.Reason,
+				"input", reActResp.Action.Input,
+			)
+
+			// Sanitize and prepare action
+			reActResp.Action.Input = sanitizeToolInput(reActResp.Action.Input)
+			a.addMessage(ctx, "assistant", fmt.Sprintf("Action: Using %s tool", reActResp.Action.Name))
+
+			// Display action details
+			u.RenderOutput(ctx, fmt.Sprintf("  Running: %s", reActResp.Action.Input), ui.Foreground(ui.ColorGreen))
+			u.RenderOutput(ctx, reActResp.Action.Reason, ui.RenderMarkdown())
+
+			// Execute action
+			output, err := a.executeAction(ctx, reActResp.Action, workDir)
+			if err != nil {
+				log.Error(err, "Error executing action")
+				return err
+			}
+
+			// Record observation
+			observation := fmt.Sprintf("Observation from %s:\n%s", reActResp.Action.Name, output)
+			a.addMessage(ctx, "system", observation)
+		}
+
+		a.CurrentIteration++
+	}
+
+	// Handle max iterations reached
+	log.Info("Max iterations reached", "iterations", a.CurrentIteration)
+	u.RenderOutput(ctx, fmt.Sprintf("\nSorry, Couldn't complete the task after %d attempts.\n", a.MaxIterations), ui.Foreground(ui.ColorRed))
+	return a.recordError(ctx, fmt.Errorf("max iterations reached"))
+}
+
+func (a *Strategy) recordError(ctx context.Context, err error) error {
+	return a.Recorder.Write(ctx, &journal.Event{
+		Timestamp: time.Now(),
+		Action:    "error",
+		Payload:   err.Error(),
+	})
+}
+
+// executeAction handles the execution of a single action
+func (a *Strategy) executeAction(ctx context.Context, action *Action, workDir string) (string, error) {
+	log := klog.FromContext(ctx)
+
+	tool := a.Tools[action.Name]
+	if tool == nil {
+		a.addMessage(ctx, "system", fmt.Sprintf("Error: Tool %s not found", action.Name))
+		log.Info("Unknown action: ", "action", action.Name)
+		return "", fmt.Errorf("unknown action: %s", action.Name)
+	}
+
+	output, err := tool(action.Input, a.Kubeconfig, workDir)
+	if err != nil {
+		return fmt.Sprintf("Error executing %q command: %v", action.Name, err), err
+	}
+	return output, nil
+}
+
+func (a *Strategy) addMessage(ctx context.Context, role, content string) error {
+	log := klog.FromContext(ctx)
+	log.Info("Tracing...")
+
+	msg := Message{
+		Role:    role,
+		Content: content,
+	}
+	a.Messages = append(a.Messages, msg)
+	a.Recorder.Write(ctx, &journal.Event{
+		Timestamp: time.Now(),
+		Action:    "trace",
+		Payload:   msg,
+	})
+
+	return nil
+}
+
+// parseReActResponse parses the LLM response into a ReActResponse struct
+// This function assumes the input contains exactly one JSON code block
+// formatted with ```json and ``` markers. The JSON block is expected to
+// contain a valid ReActResponse object.
+func parseReActResponse(input string) (*ReActResponse, error) {
+	cleaned := strings.TrimSpace(input)
+
+	const jsonBlockMarker = "```json"
+	first := strings.Index(cleaned, jsonBlockMarker)
+	last := strings.LastIndex(cleaned, "```")
+	if first == -1 || last == -1 {
+		return nil, fmt.Errorf("no JSON code block found in %q", cleaned)
+	}
+	cleaned = cleaned[first+len(jsonBlockMarker) : last]
+
+	cleaned = strings.ReplaceAll(cleaned, "\n", "")
+	cleaned = strings.TrimSpace(cleaned)
+
+	var reActResp ReActResponse
+	if err := json.Unmarshal([]byte(cleaned), &reActResp); err != nil {
+		return nil, fmt.Errorf("error parsing response %q: %w", cleaned, err)
+	}
+	return &reActResp, nil
+}
+
+func sanitizeToolInput(input string) string {
+	return strings.TrimSpace(input)
+}
